+++ NEWS.md
@@ -5,6 +5,7 @@
 * Column names for `model_performance()` and `compare_performance()` were changed to be in line with the _easystats_ naming convention: `LOGLOSS` is now `Log_loss`, `SCORE_LOG` is `Score_log` and `SCORE_SPHERICAL` is now `Score_spherical`.
 
 ## New functions
+* `r2_posterior()` for Bayesian models to obtain posterior distributions of R-squared.
 * `r2_bayes()` works with Bayesian models from `BayesFactor` ( #143 ).
 
 ## Changes to functions